delete confirmation added

+++ src/employee/components/Card.js
@@ -4,6 +4,7 @@ import {useNavigate} from "react-router-dom";
 import { bindActionCreators } from "redux";
 import { connect } from "react-redux";
 import { deleteEmployee } from "../action"; 
+import Swal from 'sweetalert2'
 const IMAGE_PLACEHOLDER = "https://cdn.pixabay.com/photo/2015/10/05/22/37/blank-profile-picture-973460__340.png"
 
 function Card({id,img,name,email,mobile,gender,deleteEmployee}) {
@@ -14,8 +15,26 @@ function Card({id,img,name,email,mobile,gender,deleteEmployee}) {
       });
   }
   const handleDeleteEmployee = (id) =>{
-    deleteEmployee(id)
-    navigate("/employee/list")
+    Swal.fire({
+      title: 'Error!',
+      text: 'Do you want to delete this employee',
+      icon: 'error',
+      showCancelButton: true,
+      confirmButtonText: 'Yes delete it!',
+    }).then((value) =>{
+      if(value.isConfirmed){
+        Swal.fire(
+          'Deleted!',
+          'Employee Delete Successfully',
+          'success'
+        ).then((value) =>{
+          if(value){
+            deleteEmployee(id)
+            navigate("/employee/list")
+          }
+        })
+      }
+    })
   }
   return (
     <div className="card employee-card">
